be-image/requirements.txt
scipy==1.3.1
numpy==1.17.1
six==1.11.0
keras==2.2.4
matplotlib==2.1.2
nltk==3.3.0
pandas==1.0.1
scikit-learn==0.21.2
bcolz==1.1.2
IPython==7.12.0
geojson==2.4.0
geopy==1.18.1
flask-restplus==0.11.0
Flask-Cors==3.0.8
requests==2.18.4
shapely==1.6.4.post2
flask-sqlalchemy==2.3.2
Pillow==5.0.0
tensorflow==1.14.0
psutil==5.6.3
stomp.py==4.1.22
theano==1.0.4
werkzeug==0.16.0